feat(auth): move init_salt to auth using strategies for persistence
