+++ src/console/src/auth/mod.rs
@@ -1,3 +1,2 @@
-mod salt;
 pub mod store;
 mod strategy_impls;
