+++ src/console/src/auth/store.rs
@@ -1,12 +1,7 @@
-use crate::auth::salt::init_salt;
 use crate::auth::strategy_impls::AuthHeap;
-use junobuild_auth::state::{
-    get_config as get_state_config, get_salt as get_state_salt, insert_salt,
-    set_config as set_store_config,
-};
+use junobuild_auth::state::{get_config as get_state_config, set_config as set_store_config};
 use junobuild_auth::types::config::AuthenticationConfig;
 use junobuild_auth::types::interface::SetAuthenticationConfig;
-use junobuild_auth::types::state::Salt;
 
 pub async fn set_config(
     proposed_config: &SetAuthenticationConfig,
@@ -16,7 +11,7 @@ pub async fn set_config(
     // If we ever need it, in the Satellite we update_alternative_origins here.
 
     if config.google.is_some() {
-        init_salt().await?;
+        junobuild_auth::state::init_salt(&AuthHeap).await?;
     }
 
     Ok(config)
@@ -24,12 +19,4 @@ pub async fn set_config(
 
 pub fn get_config() -> Option<AuthenticationConfig> {
     get_state_config(&AuthHeap)
-}
-
-pub fn set_salt(salt: &Salt) {
-    insert_salt(&AuthHeap, salt);
-}
-
-pub fn get_salt() -> Option<Salt> {
-    get_state_salt(&AuthHeap)
 }
