+++ src/console/src/certification/strategy_impls.rs
@@ -1,9 +1,7 @@
-use crate::auth::store::get_salt;
 use crate::certification::cert::update_certified_data;
 use ic_certification::HashTree;
 use junobuild_auth::state::pruned_labeled_sigs_root_hash_tree;
 use junobuild_auth::strategies::AuthCertificateStrategy;
-use junobuild_auth::types::state::Salt;
 use junobuild_storage::strategies::StorageCertificateStrategy;
 
 /// The consumer of junobuild_auth and junobuild_storage is responsible for implementing
@@ -15,10 +13,6 @@ pub struct AuthCertificate;
 impl AuthCertificateStrategy for AuthCertificate {
     fn update_certified_data(&self) {
         update_certified_data()
-    }
-
-    fn salt(&self) -> Option<Salt> {
-        get_salt()
     }
 }
