+++ src/libs/satellite/src/auth/mod.rs
@@ -1,5 +1,4 @@
 pub mod alternative_origins;
 pub mod assert;
-mod salt;
 pub mod store;
 mod strategy_impls;
