+++ src/libs/satellite/src/auth/store.rs
@@ -1,13 +1,8 @@
 use crate::auth::alternative_origins::update_alternative_origins;
-use crate::auth::salt::init_salt;
 use crate::auth::strategy_impls::AuthHeap;
-use junobuild_auth::state::{
-    get_config as get_state_config, get_salt as get_state_salt, insert_salt,
-    set_config as set_store_config,
-};
+use junobuild_auth::state::{get_config as get_state_config, set_config as set_store_config};
 use junobuild_auth::types::config::AuthenticationConfig;
 use junobuild_auth::types::interface::SetAuthenticationConfig;
-use junobuild_auth::types::state::Salt;
 
 pub async fn set_config(
     proposed_config: &SetAuthenticationConfig,
@@ -17,7 +12,7 @@ pub async fn set_config(
     update_alternative_origins(&config)?;
 
     if config.google.is_some() {
-        init_salt().await?;
+        junobuild_auth::state::init_salt(&AuthHeap).await?;
     }
 
     Ok(config)
@@ -25,12 +20,4 @@ pub async fn set_config(
 
 pub fn get_config() -> Option<AuthenticationConfig> {
     get_state_config(&AuthHeap)
-}
-
-pub fn set_salt(salt: &Salt) {
-    insert_salt(&AuthHeap, salt);
-}
-
-pub fn get_salt() -> Option<Salt> {
-    get_state_salt(&AuthHeap)
 }
